Move xpub-controller to component-submission
