+++ packages/component-submission/server/resolvers/removeSupportingFiles.test.js
@@ -1,8 +1,9 @@
 const { createTables } = require('@pubsweet/db-manager')
-const { SupportingFiles } = require('@elifesciences/xpub-controller')
 const AuditLog = require('@elifesciences/component-model-audit-log')
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
+
+const { SupportingFiles } = require('../use-cases')
 const { userData } = require('./index.test.data')
 
 const dummyStorage = {
