+++ packages/component-submission/server/resolvers/updateManuscript.js
@@ -1,6 +1,7 @@
 const config = require('config')
 const User = require('@elifesciences/component-model-user').model
-const { Manuscript } = require('@elifesciences/xpub-controller')
+
+const { Manuscript } = require('../use-cases')
 
 async function updateManuscript(_, { data }, { user }) {
   const userUuid = await User.getUuidForProfile(user)
