+++ packages/component-submission/server/resolvers/uploadManuscript.js
@@ -1,8 +1,9 @@
 const config = require('config')
 const User = require('@elifesciences/component-model-user').model
-const { Manuscript } = require('@elifesciences/xpub-controller')
 const { S3Storage, ScienceBeamApi } = require('@elifesciences/xpub-client')
 const pubsubManager = require('pubsweet-server/src/graphql/pubsub')
+
+const { Manuscript } = require('../use-cases')
 
 async function uploadManuscript(_, { file, id, fileSize }, { user }) {
   const userUuid = await User.getUuidForProfile(user)
