+++ packages/component-submission/server/services/index.js
@@ -1,0 +1,5 @@
+const Notification = require('./notification')
+
+module.exports = {
+  Notification,
+}
