+++ packages/component-submission/server/use-cases/helpers/manuscript.js
