+++ packages/component-submission/server/use-cases/manuscript.js
