--- server/xpub-controller/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@elifesciences/xpub-controller",
-  "version": "1.0.0",
-  "description": "",
-  "license": "MIT",
-  "author": "",
-  "main": "index.js",
-  "scripts": {},
-  "keywords": [],
-  "dependencies": {
-  },
-  "peerDependencies": {
-    "pubsweet-server": ">=2.0.0"
-  }
-}
